+++ src/components/Sidebar.js
@@ -3,12 +3,12 @@ import { connect } from 'redux-zero/react'
 import { useTranslation } from 'react-i18next'
 import PropTypes from 'prop-types'
 import { FiLayers } from 'react-icons/fi'
-import { BiNetworkChart, BiSelection, BiText } from 'react-icons/bi'
+import { BiNetworkChart, BiSelection } from 'react-icons/bi'
 import {
   BsCodeSlash, BsFilter, BsPencilSquare, BsSearch
 } from 'react-icons/bs'
 import { AiOutlineArrowLeft, AiOutlineArrowRight } from 'react-icons/ai'
-import { IoBuildSharp, IoGitNetwork } from 'react-icons/io5'
+import { IoGitNetwork } from 'react-icons/io5'
 import { IoMdOptions } from 'react-icons/io'
 import {
   FaFileAlt, FaFileExport, FaPaintBrush, FaRegCircle, FaRegHandPointer, FaStickyNote
@@ -22,14 +22,12 @@ import {
   IS_ELEMENT_SELECTION_VISIBLE,
   IS_ELEMENTS_FILTER_VISIBLE,
   IS_EXPORT_VISIBLE,
-  IS_FREE_TEXT_SEARCH_VISIBLE,
   IS_GRAPH_OPTIONS_VISIBLE,
   IS_GRAPHS_VISIBLE,
   IS_NEIGHBOURHOOD_VISIBLE,
   IS_NOTES_VISIBLE,
   IS_SEARCH_VISIBLE,
   IS_SHORTEST_PATH_VISIBLE,
-  IS_STRUCTURED_SEARCH_VISIBLE,
   IS_STYLING_VISIBLE,
   IS_SYNONIMS_VISIBLE,
   SIDEBAR_VIEW_BOUNDING_BOX,
@@ -39,18 +37,15 @@ import {
   SIDEBAR_VIEW_ELEMENTS_SELECTION,
   SIDEBAR_VIEW_ENTRY_SEARCH,
   SIDEBAR_VIEW_EXPORT,
-  SIDEBAR_VIEW_FREE_TEXT_SEARCH,
   SIDEBAR_VIEW_GRAPH_OPTIONS,
   SIDEBAR_VIEW_GRAPHS,
   SIDEBAR_VIEW_NEIGHBOURHOOD,
   SIDEBAR_VIEW_NOTES,
   SIDEBAR_VIEW_SHORTEST_PATH,
-  SIDEBAR_VIEW_STRUCTURED_SEARCH,
   SIDEBAR_VIEW_STYLING,
   SIDEBAR_VIEW_SYNONYMS
 } from '../constants/views'
 import NetworkGraphList from './NetworkGraphList'
-import FreeTextSearch from './FreeTextSearch'
 import NodeNeighbourhood from './NodeNeighbourhood'
 import ElementsSelection from './ElementsSelection'
 import ExportSettings from './ExportSettings'
@@ -60,7 +55,6 @@ import ElementsFilter from './ElementsFilter'
 import CustomQuery from './CustomQuery'
 import EditOntology from './EditOntology'
 import NetworkStyling from './NetworkStyling'
-import StructuredSearch from './StructuredSearch'
 import NotesList from './NotesList'
 import EntrySearch from './EntrySearch'
 import SynonymsList from './SynonymsList'
@@ -73,14 +67,12 @@ import {
   ROUTE_ELEMENTS_FILTER,
   ROUTE_ELEMENTS_SELECTION,
   ROUTE_EXPORT,
-  ROUTE_FREE_TEXT_SEARCH,
   ROUTE_NETWORK_GRAPH_OPTIONS,
   ROUTE_NETWORK_GRAPHS,
   ROUTE_NODE_NEIGHBOURHOOD,
   ROUTE_NOTES,
   ROUTE_SEARCH,
   ROUTE_SHORTEST_PATH,
-  ROUTE_STRUCTURED_SEARCH,
   ROUTE_STYLING,
   ROUTE_SYNONYMS,
 } from '../constants/routes'
@@ -117,30 +109,6 @@ const Sidebar = ({
       component: <NetworkGraphOptions />,
       label: SIDEBAR_VIEW_GRAPH_OPTIONS,
       route: ROUTE_NETWORK_GRAPH_OPTIONS
-    },
-    [SIDEBAR_VIEW_FREE_TEXT_SEARCH]: {
-      icon: (
-        <>
-          <BsSearch />
-          <BiText />
-        </>
-      ),
-      isVisible: IS_FREE_TEXT_SEARCH_VISIBLE,
-      component: <FreeTextSearch />,
-      label: SIDEBAR_VIEW_FREE_TEXT_SEARCH,
-      route: ROUTE_FREE_TEXT_SEARCH
-    },
-    [SIDEBAR_VIEW_STRUCTURED_SEARCH]: {
-      icon: (
-        <>
-          <BsSearch />
-          <IoBuildSharp />
-        </>
-      ),
-      isVisible: IS_STRUCTURED_SEARCH_VISIBLE,
-      component: <StructuredSearch />,
-      label: SIDEBAR_VIEW_STRUCTURED_SEARCH,
-      route: ROUTE_STRUCTURED_SEARCH
     },
     [SIDEBAR_VIEW_ELEMENTS_SELECTION]: {
       icon: (